UML-3606 Access for all and add lpa journey specific service classes moved to a subfolder
