+++ service-api/app/features/context/Integration/LpaContext.php
@@ -11,9 +11,9 @@ use App\Exception\BadRequestException;
 use App\Exception\NotFoundException;
 use App\Service\ActorCodes\ActorCodeService;
 use App\Service\Log\RequestTracing;
-use App\Service\Lpa\AccessForAllLpaService;
-use App\Service\Lpa\AddAccessForAllLpa;
-use App\Service\Lpa\AddLpa;
+use App\Service\Lpa\AccessForAll\AccessForAllLpaService;
+use App\Service\Lpa\AccessForAll\AddAccessForAllLpa;
+use App\Service\Lpa\AddLpa\AddLpa;
 use App\Service\Lpa\GetInstructionsAndPreferencesImages;
 use App\Service\Lpa\RemoveLpa;
 use App\Service\Lpa\SiriusLpaManager;
