+++ service-api/app/src/App/src/Handler/AccessForAllLpaConfirmationHandler.php
@@ -7,9 +7,9 @@ namespace App\Handler;
 use App\Exception\BadRequestException;
 use App\Service\Features\FeatureEnabled;
 use App\Service\Log\EventCodes;
-use App\Service\Lpa\AddAccessForAllLpa;
+use App\Service\Lpa\AccessForAll\AccessForAllLpaService;
+use App\Service\Lpa\AccessForAll\AddAccessForAllLpa;
 use App\Service\Lpa\CheckLpaCleansed;
-use App\Service\Lpa\AccessForAllLpaService;
 use Exception;
 use Laminas\Diactoros\Response\EmptyResponse;
 use Psr\Http\Message\ResponseInterface;
