+++ service-api/app/src/App/src/Handler/AccessForAllLpaValidationHandler.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace App\Handler;
 
 use App\Exception\BadRequestException;
-use App\Service\Lpa\AddAccessForAllLpa;
+use App\Service\Lpa\AccessForAll\AddAccessForAllLpa;
 use Exception;
 use Laminas\Diactoros\Response\JsonResponse;
 use Psr\Http\Message\{ResponseInterface, ServerRequestInterface};
