+++ service-api/app/test/AppTest/Service/Lpa/AddAccessForAllLpaTest.php
@@ -8,10 +8,10 @@ use App\DataAccess\Repository\Response\Lpa;
 use App\Exception\BadRequestException;
 use App\Exception\NotFoundException;
 use App\Service\Features\FeatureEnabled;
-use App\Service\Lpa\AccessForAllLpaService;
-use App\Service\Lpa\AddAccessForAllLpa;
+use App\Service\Lpa\AccessForAll\AccessForAllLpaService;
+use App\Service\Lpa\AccessForAll\AddAccessForAllLpa;
+use App\Service\Lpa\AddLpa\LpaAlreadyAdded;
 use App\Service\Lpa\FindActorInLpa;
-use App\Service\Lpa\LpaAlreadyAdded;
 use App\Service\Lpa\LpaManagerInterface;
 use App\Service\Lpa\RestrictSendingLpaForCleansing;
 use App\Service\Lpa\ValidateAccessForAllLpaRequirements;
