+++ service-api/app/test/AppTest/Service/Lpa/AddLpaTest.php
@@ -7,8 +7,8 @@ namespace AppTest\Service\Lpa;
 use App\Exception\BadRequestException;
 use App\Exception\NotFoundException;
 use App\Service\ActorCodes\ActorCodeService;
-use App\Service\Lpa\AddLpa;
-use App\Service\Lpa\LpaAlreadyAdded;
+use App\Service\Lpa\AddLpa\AddLpa;
+use App\Service\Lpa\AddLpa\LpaAlreadyAdded;
 use Fig\Http\Message\StatusCodeInterface;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\ExpectationFailedException;
